fix(ExecutionError): move Metadata into own model ExecutionErrorMetadata.
